Missed 2 functions; and did `renderMatrix`.
